+++ src/main/kotlin/pl/denisolek/panel/schema/DTO/type/SchemaDetailsDTO.kt
@@ -1,7 +1,7 @@
 package pl.denisolek.panel.schema.DTO.type
 
 data class SchemaDetailsDTO(
-        var width: Int,
-        var heigth: Int,
+        var width: Float,
+        var heigth: Float,
         var rotation: Float
 )
